refactor: PostItemCommentを作成
